requirements.txt
keras==2.9.0
matplotlib==3.5.3
numpy==1.23.0
pandas==1.4.3
Pillow==9.2.0
scikit_image==0.19.2
scikit_learn==1.1.2
skimage==0.0
tensorflow==2.9.1
tensorflow_gpu==2.9.1
argparse==1.4.0